requirements.txt
pandas>=1.1.0
xlrd>=1.0.0
beautifulsoup4>=4.9.1
pdfminer.six>=20200726
confuse>=1.3.0